native-port: reimplement Metrowerks inline PPC asm kernels in portable C++ (task #6)
Replace hand-written Gekko paired-single/vector asm in the engine TUs with
behaviorally-equivalent portable C++ (no GameCube/PPC), and fix the pointer-cast
"loses precision" errors in the same files (skipped by the earlier cast campaign
because they contain the token `asm`).

MarioUtil/MathUtil.cpp
  - MsVECMag2: returns sqrtf(x*x+y*y+z*z) (vector magnitude).
  - MsVECNormalize: scales v1 by 1/sqrt(|v1|^2) into v2 (normalize).

JSystem/J3D/J3DGraphBase/J3DTransform.cpp
  - J3DPSCalcInverseTranspose: cofactor/det inverse-transpose of the upper-left
    3x3, stored as ROMtx; returns false when singular.
  - J3DScaleNrmMtx33: per-component row scale mtx[i][j] *= scl[j].
  - J3DMtxProjConcat: result(3x4) = param_1(3x4) x param_2(4x4) projection concat.
  - J3DPSMtx33Copy / J3DPSMtx33CopyFrom34: copy upper-left 3x3.
  - J3DPSMtxArrayCopy: copy an array of 3x4 matrices.
  - J3DPSMtxArrayConcat: dst[k] = fst x snd[k] affine concat over an array.
  - J3DMTXConcatArrayIndexedSrc: dst[k] = mat1 x mat2[index[k]] affine concat.

JSystem/J3D/J3DGraphAnimator/J3DAnimation.cpp
  - J3DHermiteInterpolationS: register-faithful translation of the paired-single
    Hermite interpolation kernel (verified bit-equivalent to the asm; the
    commented reference C in the source is the author's non-matching guess).

JSystem/JUtility/JUTException.cpp
  - getFpscr: no host Gekko FPSCR; returns 0. Fix createFB pointer casts.

JSystem/JKernel/JKRHeap.cpp (+ JKRHeap.hpp)
  - dispose / dispose_subroutine: widen address params u32 -> uintptr_t so heap
    addresses survive the pointer round-trip on 64-bit (a real correctness fix,
    not just the cast warning).

Compile-clean sweep over reference/sms/src/*.cpp (native flags): 386 -> 388.
J3DTransform.cpp and MathUtil.cpp now fully native-clean; J3DAnimation already
compiled clean (asm was MWERKS-gated) and now actually works natively. JKRHeap
and JUTException retain only unrelated (non-asm/non-cast) errors. No regressions.

+++ include/JSystem/JKernel/JKRHeap.hpp
@@ -99,7 +99,7 @@ public:
 	void* getMaxFreeBlock();
 	u32 getMaxAllocatableSize(int alignment);
 	JKRHeap* find(void*) const;
-	void dispose_subroutine(u32 begin, u32 end);
+	void dispose_subroutine(uintptr_t begin, uintptr_t end);
 	bool dispose(void*, u32);
 	void dispose(void*, void*);
 	void dispose();
