+++ src/JSystem/JKernel/JKRHeap.cpp
@@ -144,7 +144,7 @@ JKRHeap* JKRHeap::find(void* memory) const
 	return nullptr;
 }
 
-void JKRHeap::dispose_subroutine(u32 begin, u32 end)
+void JKRHeap::dispose_subroutine(uintptr_t begin, uintptr_t end)
 {
 	JSUListIterator<JKRDisposer> last_iterator;
 	JSUListIterator<JKRDisposer> next_iterator;
@@ -175,15 +175,15 @@ void JKRHeap::dispose_subroutine(u32 begin, u32 end)
 
 bool JKRHeap::dispose(void* memory, u32 size)
 {
-	u32 begin = (u32)memory;
-	u32 end   = (u32)memory + size;
+	uintptr_t begin = (uintptr_t)memory;
+	uintptr_t end   = (uintptr_t)memory + size;
 	dispose_subroutine(begin, end);
 	return false;
 }
 
 void JKRHeap::dispose(void* begin, void* end)
 {
-	dispose_subroutine((u32)begin, (u32)end);
+	dispose_subroutine((uintptr_t)begin, (uintptr_t)end);
 }
 
 void JKRHeap::dispose()
